test: resources and snapshot for static using test #372
