+++ resources/c-sharp/static-using/ProgramFile.cs
@@ -1,0 +1,14 @@
+using static StaticUsedNamespace.ClassA;
+using NormalUsedNamespace;
+
+namespace ProgramNamespace
+{
+    class ProgramClass
+    {
+        static void Main()
+        {
+            MyMethod();
+            ClassB.MyMethod();
+        }
+    }
+}
